update  order details page
