--- src/components/Faq/Faq.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Faq() {
-  return <div></div>;
-}
-
-export default Faq;
