+++ src/components/FaqDetails/FaqDetails.scss
