+++ src/components/RewardsDetails/RewardsDetails.scss
@@ -33,11 +33,13 @@
     align-items: centerdddd;
     gap: 1em;
     padding-bottom: 4em;
+    width: 100%;
   }
 
   &__paragraph {
     padding-top: 2em;
   }
+
   &__prize {
     width: 100%;
     display: flex;
