+++ src/components/Toggle/Toggle.scss
@@ -4,7 +4,9 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
+  width: 100%;
   @include background-style;
+
   &__btn {
     background-color: var(--icon-color);
     border: 1px solid var(--icon-color);
