+++ src/pages/Cart/Cart.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import CartDetails from "../../components/CartDetails/CartDetails";
 
 function Cart() {
