+++ src/pages/Faq/Faq.jsx
@@ -1,58 +1,13 @@
-import Navigation from "../../components/Navigation/Navigation.jsx";
-import "../Faq/Faq.scss";
-import React, { useState } from "react";
+import Navigation from "../../components/Navigation/Navigation";
+import FaqDetails from "../../components/FaqDetails/FaqDetails";
 
 function Faq() {
-  const [selected, setSelected] = useState(null);
-
-  const toggle = (i) => {
-    if (selected === i) {
-      return setSelected(null);
-    }
-
-    setSelected(i);
-  };
   return (
-    <div className="wrapper">
+    <div>
       <Navigation />
-      <h1>FAQ</h1>
-      <h4>Ordering</h4>
-      <div className="accordion">
-        {data.map((item, i) => (
-          <div className="item" key={item.id}>
-            <div className="title" onClick={() => toggle(i)}>
-              <p className="body-bold">{item.question}</p>
-              <span>{selected === i ? "-" : "+"}</span>
-            </div>
-
-            {selected === i && <div className="content">{item.answer}</div>}
-          </div>
-        ))}
-      </div>
-      <Faq2 />
+      <FaqDetails />
     </div>
   );
 }
-
-const data = [
-  {
-    id: 1,
-    question: "How can I place an order?",
-    answer:
-      "To place an order, you have many options. Drop by the shop to order in-person during our regular hours, order ahead for pick-up on Ritual, our app or for delivery on Uber Eats and Doordash.",
-  },
-  {
-    id: 2,
-    question: "Can I order over the phone?",
-    answer:
-      "Phone orders are not possible at this time as we cannot process a transaction for your order via text or phone.",
-  },
-  {
-    id: 3,
-    question: "Do you offer catering?",
-    answer:
-      "Yes! Burger Drops specializes in both on-site private catering and large group orders for your office or next celebration.",
-  },
-];
 
 export default Faq;
