+++ src/pages/OrderConfirmation/OrderConfirmation.scss
@@ -1,2 +1,9 @@
-@use "../../styles/global";
-@use "../../styles/mixins";
+@use "../../styles/mixins" as *;
+
+.order__text {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  @include background-style;
+}
